+++ demos/parse-server-migration/src/cloud/main.ts
@@ -6,12 +6,12 @@ import './generated/solApi';
 import { requestMessage } from '../auth/authService';
 
 Parse.Cloud.define('requestMessage', async ({ params }: any) => {
-  const { address, chain, network } = params;
+  const { address, chain, networkType } = params;
 
   const message = await requestMessage({
     address,
     chain,
-    network,
+    networkType,
   });
 
   return { message };