+++ demos/parse-server/src/auth/authController.ts
@@ -3,12 +3,12 @@ import { NextFunction, Request, Response } from 'express';
 
 export async function request(req: Request, res: Response, next: NextFunction) {
   try {
-    const { address, chain, network } = req.body;
+    const { address, chain, networkType } = req.body;
 
     const message = await requestMessage({
       address,
       chain,
-      network,
+      networkType,
     });
 
     res.status(200).json({ message });