+++ packages/integration/mockRequests/config.ts
@@ -1,4 +1,0 @@
-export const MOCK_API_KEY = 'test-api-key';
-export const EVM_API_ROOT = 'https://deep-index.moralis.io/api/v2';
-export const SOL_API_ROOT = 'https://solana-gateway.moralis.io';
-export const STREAM_API_ROOT = 'https://streams-api.aws-prod-streams-master-1.moralis.io';