+++ packages/streams/src/methods/getAll.ts
@@ -1,32 +1,26 @@
-import { BASE_URL } from '../MoralisStreams';
-import { PaginatedEndpointResolver } from '@moralisweb3/api-utils';
-import MoralisCore, { MoralisStreamError, StreamErrorCode } from '@moralisweb3/core';
-import { getStreamsEvm } from '../resolvers';
-export enum StreamNetwork {
-  EVM = 'evm',
-}
+import { Endpoints } from '@moralisweb3/api-utils';
+import { getStreamsEvm, GetStreamsEvmParams } from '../resolvers';
+import { StreamNetwork } from '../utils/StreamNetwork';
+import { IncorrectNetworkError } from '../utils/IncorrectNetworkError';
 
-export interface GetStreamsEvmOptions {
-  network: 'evm';
-  limit: number;
+export interface GetStreamsEvmOptions extends GetStreamsEvmParams {
+  networkType?: 'evm';
 }
 
 export type GetStreamsOptions = GetStreamsEvmOptions;
 
-const makeGetStreamEvm = (core: MoralisCore, { network, ...options }: GetStreamsEvmOptions) => {
-  return PaginatedEndpointResolver.create(core, BASE_URL, getStreamsEvm).fetch(options);
-};
+export const makeGetStreams = (endpoints: Endpoints) => {
+  const evmFetcher = endpoints.createPaginatedFetcher(getStreamsEvm);
 
-export const getStreams = (core: MoralisCore) => (options: GetStreamsOptions) => {
-  switch (options.network) {
-    case StreamNetwork.EVM:
-      return makeGetStreamEvm(core, options);
-    default:
-      throw new MoralisStreamError({
-        code: StreamErrorCode.INCORRECT_NETWORK,
-        message: `Incorrect network provided. Got "${options.network}", Valid values are: ${Object.values(StreamNetwork)
-          .map((value) => `"${value}"`)
-          .join(', ')}`,
-      });
-  }
+  return ({ networkType, ...options }: GetStreamsOptions) => {
+    switch (networkType) {
+      case StreamNetwork.EVM:
+        return evmFetcher({ ...options });
+      default:
+        if (networkType === undefined) {
+          return evmFetcher({ ...options });
+        }
+        throw new IncorrectNetworkError(networkType);
+    }
+  };
 };