+++ packages/solApi/CHANGELOG.md
@@ -1,5 +1,146 @@
 # @moralisweb3/sol-api
 
+## 2.6.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.6.5
+  - @moralisweb3/api-utils@2.6.5
+  - @moralisweb3/sol-utils@2.6.5
+
+## 2.6.4
+
+### Patch Changes
+
+- Updated dependencies [[`77eb29bf`](https://github.com/MoralisWeb3/Moralis-JS-SDK/commit/77eb29bf7d920e40a74028ed989a5b21f2dc9ec4)]:
+  - @moralisweb3/api-utils@2.6.4
+  - @moralisweb3/core@2.6.4
+  - @moralisweb3/sol-utils@2.6.4
+
+## 2.6.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.6.3
+  - @moralisweb3/api-utils@2.6.3
+  - @moralisweb3/sol-utils@2.6.3
+
+## 2.6.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.6.2
+  - @moralisweb3/api-utils@2.6.2
+  - @moralisweb3/sol-utils@2.6.2
+
+## 2.6.1
+
+### Patch Changes
+
+- Updated dependencies [[`bfea2ba5`](https://github.com/MoralisWeb3/Moralis-JS-SDK/commit/bfea2ba51f0616d097aa7d3251af516d632eb502)]:
+  - @moralisweb3/sol-utils@2.6.1
+  - @moralisweb3/core@2.6.1
+  - @moralisweb3/api-utils@2.6.1
+
+## 2.6.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.6.0
+  - @moralisweb3/api-utils@2.6.0
+  - @moralisweb3/sol-utils@2.6.0
+
+## 2.5.8
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.5.8
+  - @moralisweb3/api-utils@2.5.8
+  - @moralisweb3/sol-utils@2.5.8
+
+## 2.5.7
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.5.7
+  - @moralisweb3/api-utils@2.5.7
+  - @moralisweb3/sol-utils@2.5.7
+
+## 2.5.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.5.6
+  - @moralisweb3/api-utils@2.5.6
+  - @moralisweb3/sol-utils@2.5.6
+
+## 2.5.5
+
+### Patch Changes
+
+- [#699](https://github.com/MoralisWeb3/Moralis-JS-SDK/pull/699) [`d79800ad`](https://github.com/MoralisWeb3/Moralis-JS-SDK/commit/d79800ade15a715107de2c7e3e28c4be4c1d7cf6) Thanks [@ErnoW](https://github.com/ErnoW)! - Fix parsing of API error messages, now the MoralisError will show the `message` that is returned from the api.
+
+- Updated dependencies [[`d79800ad`](https://github.com/MoralisWeb3/Moralis-JS-SDK/commit/d79800ade15a715107de2c7e3e28c4be4c1d7cf6)]:
+  - @moralisweb3/core@2.5.5
+  - @moralisweb3/api-utils@2.5.5
+  - @moralisweb3/sol-utils@2.5.5
+
+## 2.5.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.5.4
+  - @moralisweb3/api-utils@2.5.4
+  - @moralisweb3/sol-utils@2.5.4
+
+## 2.5.3
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.5.3
+  - @moralisweb3/api-utils@2.5.3
+  - @moralisweb3/sol-utils@2.5.3
+
+## 2.5.2
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.5.2
+  - @moralisweb3/api-utils@2.5.2
+  - @moralisweb3/sol-utils@2.5.2
+
+## 2.5.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @moralisweb3/core@2.5.1
+  - @moralisweb3/api-utils@2.5.1
+  - @moralisweb3/sol-utils@2.5.1
+
+## 2.5.0
+
+### Minor Changes
+
+- [#670](https://github.com/MoralisWeb3/Moralis-JS-SDK/pull/670) [`f6696b86`](https://github.com/MoralisWeb3/Moralis-JS-SDK/commit/f6696b864904044a84746b6ad805724b00e6a88d) Thanks [@3scava1i3r](https://github.com/3scava1i3r)! - Added SolApi.token.getTokenPrice
+
+### Patch Changes
+
+- Updated dependencies [[`c719cc2f`](https://github.com/MoralisWeb3/Moralis-JS-SDK/commit/c719cc2fe0b6bb31093bbaa8a6907b6697b7b129)]:
+  - @moralisweb3/core@2.5.0
+  - @moralisweb3/api-utils@2.5.0
+  - @moralisweb3/sol-utils@2.5.0
+
 ## 2.4.0
 
 ### Patch Changes